move Spin1Gen and Spin2Gen to their own file
